\name{SCM}
\alias{SCM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Sample Covariance Matrix
}
\description{
This function creates a sample covariance matrix from an EEG trial matrix.
}
\usage{
SCM(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a nxt numeric matrix, where n indicates the number of recording channels and t indicates the time points.
}
}

\value{
It returns a sample covariance matrix.
}
\references{
Barachant, Alexandre, Stéphane Bonnet, Marco Congedo e Christian Jutten (2012). "Multiclass brain-computer interface classification by Riemannian geometry". In: IEEE Transactions on Biomedical Engineering. doi: 10.1109/TBME.2011.2172210.
}
\author{
Laura Masiero, email: laura.masiero.10@gmail.com
}



\keyword{ ~SCM }

